requirements.txt
streamlit==1.55.0
yfinance==1.2.0
pandas==2.3.3
plotly==6.6.0
ta==0.11.0
vaderSentiment==3.3.2
newsapi-python==0.2.7
python-dotenv==1.2.2
requests==2.32.5
numpy==2.4.3
